@@ -29,9 +29,9 @@
 
 
         /**
-         * Clean after the component is destroyed.
+         * Clean after the component is unmounted.
          */
-        destroyed() {
+        unmounted() {
             clearTimeout(this.timeout);
         },
 
