@@ -7,26 +7,11 @@ use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Queue\QueueManager;
 use Illuminate\Queue\Worker;
 use Illuminate\Queue\WorkerOptions;
-use Orchestra\Testbench\Concerns\CreatesApplication;
 
-$appLoader = new class
-{
-    use CreatesApplication;
-
-    /**
-     * Define environment setup.
-     *
-     * @param  \Illuminate\Foundation\Application  $app
-     * @return void
-     */
-    protected function getEnvironmentSetUp($app)
-    {
-        // Define your environment setup.
-    }
-};
+use function Orchestra\Testbench\container;
 
 // Configure the application...
-$app = $appLoader->createApplication();
+$app = container()->createApplication();
 $app->register(Laravel\Horizon\HorizonServiceProvider::class);
 $app->make('config')->set('queue.default', 'redis');
 
