@@ -1,8 +1,8 @@
-import vue2 from '@vitejs/plugin-vue2';
+import vue from '@vitejs/plugin-vue';
 
 /** @type {import('vite').UserConfig} */
 export default {
-    plugins: [vue2()],
+    plugins: [vue()],
     build: {
         assetsDir: '',
         rollupOptions: {
